Refactored the get game function into the get_prediction_data file.
